Apply changes suggested by static analyzer SonarQube by ide labs was used

+++ include/i8080_disassembler.h
@@ -2,6 +2,7 @@
 #define __I8080_DISASSEMBLER_H__
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "bytestream.h"
 
 #define I8080_LITTLE_ENDIAN  true
@@ -51,7 +52,7 @@ typedef struct {
 i8080_line_t i8080_instruction_toString(i8080_instruction_t instruction, bool generateComments);
 
 // returns a struct describing the instruction found at the specified address in the memory 
-i8080_instruction_t disassemble_instruction(const uint8_t* mem , unsigned int addr);
+i8080_instruction_t disassemble_instruction(const uint8_t* mem , uint16_t addr);
 
 // returns a string containing the text disassembly of the given program
 void disassemble_program(const bytestream_t program, FILE* ofp);
